Add services, controllers and repositories for trip domain
- Implement TripService (create, list, detail, update, delete
- Implement TripDayService (add, update, delete days)
- Implement ActivityService (add, update, delete activities)
- Implement TripMemberService (add, update role, remove members)
- Add TripAuthorizationService (membership, editor/owner, owner validation)
- Split controllers by domain (Trip, TripDay, Activity, TripMember)
- Add ActivityRepository, TripDayRepository, UserTripRepository
- Fix TripRepository to use JpaRepository<Trip, Long>

+++ src/main/java/com/mcesnik/planner_backend/repository/TripRepository.java
@@ -2,10 +2,9 @@ package com.mcesnik.planner_backend.repository;
 
 import com.mcesnik.planner_backend.model.Trip;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TripRepository extends JpaRepository<Trip, Integer> {
+public interface TripRepository extends JpaRepository<Trip, Long> {
 
 }
